@@ -13,11 +13,8 @@ class Teachers::ProgressReports::Concepts::ConceptsController < Teachers::Progre
   end
 
   private def json_payload
-    concepts_json = current_user.all_classrooms_cache(key: 'teachers.progress_reports.concepts.concepts') do
-      concepts_as_json
-    end
     {
-      concepts: concepts_json,
+      concepts: concepts_as_json,
       student: {name: student.name}
     }
   end
