@@ -91,8 +91,8 @@ class Teachers::ProgressReports::DiagnosticReportsController < Teachers::Progres
 
       if pre_test && pre_test_activity_session
         concept_results = {
-          pre: { questions: format_concept_results(pre_test_activity_session, pre_test_activity_session.old_concept_results.order("(metadata->>'questionNumber')::int")) },
-          post: { questions: format_concept_results(activity_session, activity_session.old_concept_results.order("(metadata->>'questionNumber')::int")) }
+          pre: { questions: format_concept_results(pre_test_activity_session, pre_test_activity_session.old_concept_results.order(Arel.sql("(metadata->>'questionNumber')::int"))) },
+          post: { questions: format_concept_results(activity_session, activity_session.old_concept_results.order(Arel.sql("(metadata->>'questionNumber')::int"))) }
         }
         formatted_skills = skills.map do |skill|
           {
@@ -102,7 +102,7 @@ class Teachers::ProgressReports::DiagnosticReportsController < Teachers::Progres
         end
         skill_results = { skills: formatted_skills.uniq { |formatted_skill| formatted_skill[:pre][:skill] } }
       else
-        concept_results = { questions: format_concept_results(activity_session, activity_session.old_concept_results.order("(metadata->>'questionNumber')::int")) }
+        concept_results = { questions: format_concept_results(activity_session, activity_session.old_concept_results.order(Arel.sql("(metadata->>'questionNumber')::int"))) }
         skill_results = { skills: skills.map { |skill| data_for_skill_by_activity_session(activity_session.old_concept_results, skill) }.uniq { |formatted_skill| formatted_skill[:skill] } }
       end
       { concept_results: concept_results, skill_results: skill_results, name: student.name }
