@@ -7,7 +7,8 @@ import { Events } from '../modules/analytics'
 
 export const getActivity = (sessionID: string, activityUID: string) => {
   return async (dispatch: Function) => {
-    dispatch(TrackAnalyticsEvent(Events.COMPREHENSION_ACTIVITY_STARTED, {
+
+    dispatch(TrackAnalyticsEvent(Events.EVIDENCE_ACTIVITY_STARTED, {
       activityID: activityUID,
       sessionID: sessionID
     }));
