@@ -101,7 +101,8 @@ export {
   titleCase,
   onMobile,
   fillInBlankInputLabel,
-  getIconForActivityClassification
+  getIconForActivityClassification,
+  isTrackableStudentEvent
 } from './libs/index'
 
 export {
@@ -180,5 +181,10 @@ export {
   PRIVATE_FLAG,
   NOT_APPLICABLE
 } from './utils/constants'
+
+export {
+  UserIdsForEvent,
+  Question,
+  QuestionObject } from './interfaces'
 
 export { DefaultReactQueryClient } from './utils/defaultReactQueryClient'
