@@ -1,11 +1,13 @@
 import * as React from 'react';
+import { renderToString } from 'react-dom/server'
 
-import { Activity, ActivityClassification, Topic, } from './interfaces'
+import { Activity, Topic, } from './interfaces'
 import { stringifyLowerLevelTopics, AVERAGE_FONT_WIDTH, } from './shared'
 
 import { imageTagForClassification, } from '../../assignmentFlowConstants'
 import NumberSuffixBuilder from '../../../modules/numberSuffixBuilder'
 import useWindowSize from '../../../../../Shared/hooks/useWindowSize'
+import { Tooltip, } from '../../../../../Shared/index'
 
 const smallWhiteCheckSrc = `${process.env.CDN_URL}/images/shared/check-small-white.svg`
 const expandSrc = `${process.env.CDN_URL}/images/shared/expand.svg`
@@ -27,8 +29,15 @@ const readabilityCopy = "Quill recommends using activities where the text readab
 
 const readabilityContent = (activity) => (
   <div>
-    {activity.minimum_grade_level ? <span>Grade Range: {NumberSuffixBuilder(activity.minimum_grade_level)}-{NumberSuffixBuilder(activity.maximum_grade_level)}<br /></span> : null}
-    <span>Text Readability Level: {activity.readability_grade_level}</span>
+    {activity.minimum_grade_level ? (<span>
+      <span className="grade-range">{NumberSuffixBuilder(activity.minimum_grade_level)}-{NumberSuffixBuilder(activity.maximum_grade_level)}</span>
+      Grade Range
+      <br />
+    </span>) : null}
+    <span>
+      <span className="grade-range">{activity.readability_grade_level}</span>
+      <span>Text Readability Level</span>
+    </span>
     <br />
     <br />
     <span>{readabilityCopy}</span>
@@ -59,18 +68,13 @@ interface ActivityRowProps {
 // we can't rely on pure CSS for this because the max width is so dependent on the presence and length of other elements in the row
 const calculateMaxAllowedLengthForTopicSection = ({
   activity_classification,
-  activity_category_name,
-  readability_grade_level,
-  standard_level_name
 }): number => {
   let maxAllowedLength = 0
   const container = document.getElementsByClassName('activity-table-container')[0]
   if (container) {
-    maxAllowedLength = container.offsetWidth - 152 // 128 is the amount of padding in total that makes up the difference between the width of the container and the width of the second line
+    maxAllowedLength = container.offsetWidth - 128 // 128 is the amount of padding in total that makes up the difference between the width of the container and the width of the second line
     maxAllowedLength -= activity_classification && activity_classification.alias ? (activity_classification.alias.length * AVERAGE_FONT_WIDTH) + IMAGE_WIDTH + MARGIN : 0
-    maxAllowedLength -= activity_category_name ? (activity_category_name.length * AVERAGE_FONT_WIDTH) + IMAGE_WIDTH + MARGIN : 0
-    maxAllowedLength -= readability_grade_level ? (`Readability: Grades ${readability_grade_level}`.length * AVERAGE_FONT_WIDTH) + IMAGE_WIDTH : 0 // the readability section is the only section that does not have a margin
-    maxAllowedLength -= standard_level_name ? (standard_level_name.length * AVERAGE_FONT_WIDTH) + IMAGE_WIDTH + MARGIN : 0
+    maxAllowedLength -= 374 // 374 is the standardized width for the ccss grade + grade levels
   }
   return maxAllowedLength
 }
@@ -217,9 +221,7 @@ const ActivityRowExpandedSection = ({ activity, isExpanded}: { activity: Activit
   )
 }
 
-const ActivityRowTooltip = ({ activity, showTooltip}: { activity: Activity, showTooltip: boolean }) => {
-  if (!showTooltip) { return <span />}
-
+const ActivityRowTooltip = ({ activity, }: { activity: Activity }) => {
   const descriptionLine = activity.description && (<div className="tooltip-line">
     <img alt="Information icon" src={informationSrc} />
     <span>{activity.description}</span>
@@ -238,30 +240,28 @@ const ActivityRowTooltip = ({ activity, showTooltip}: { activity: Activity, show
   )
 
   return (
-    <div className="activity-row-tooltip">
+    <React.Fragment>
       {readabilityLine}
       {descriptionLine}
       {contentPartnerLines}
-    </div>
+    </React.Fragment>
   )
 }
 
 const ActivityRow = ({ activity, isSelected, toggleActivitySelection, showCheckbox, showRemoveButton, isFirst, setShowSnackbar, saveActivity, unsaveActivity, savedActivityIds, gradeLevelFilters, }: ActivityRowProps) => {
   const size = useWindowSize();
   const [isExpanded, setIsExpanded] = React.useState(false)
-  const [showTooltip, setShowTooltip] = React.useState(false)
 
   function toggleIsExpanded() { setIsExpanded(!isExpanded) }
-  function toggleShowTooltip() { setShowTooltip(!showTooltip)}
-  function removeActivity() {
-    toggleActivitySelection(activity, isSelected)
-    setShowSnackbar && setShowSnackbar(true)
-  }
-
   const { activity_classification, name, activity_category_name, standard_level_name, anonymous_path, readability_grade_level, topics, id, minimum_grade_level, maximum_grade_level, description, content_partners, } = activity
 
   function handleClickSaveButton() { saveActivity(id) }
   function handleClickSavedButton() { unsaveActivity(id) }
+
+  function removeActivity() {
+    toggleActivitySelection(activity, isSelected)
+    setShowSnackbar && setShowSnackbar(true)
+  }
 
   const expandImgAltText = `Arrow pointing ${isExpanded ? 'up' : 'down'}`
 
@@ -289,7 +289,7 @@ const ActivityRow = ({ activity, isSelected, toggleActivitySelection, showCheckb
         <ActivityRowTopics
           hasConcept={!!activity_category_name}
           inExpandedView={isExpanded}
-          maxAllowedLength={calculateMaxAllowedLengthForTopicSection({ activity_classification, activity_category_name, readability_grade_level, standard_level_name})}
+          maxAllowedLength={calculateMaxAllowedLengthForTopicSection({ activity_classification, })}
           onTertiaryLine={true}
           topics={topics}
         />
@@ -297,13 +297,17 @@ const ActivityRow = ({ activity, isSelected, toggleActivitySelection, showCheckb
     )
   }
 
+  const tooltipContent = renderToString(<ActivityRowTooltip activity={activity} />)
+
   return (
     <section className={`activity-row ${expandClassName} ${isSelectedClassName} ${isFirstClassName} ${expandedButEmptyClassName}`}>
-      <ActivityRowTooltip activity={activity} showTooltip={showTooltip} />
       <div className="first-line">
         <div className="name-and-checkbox-wrapper">
           {showCheckbox && <ActivityRowCheckbox activity={activity} isSelected={isSelected} toggleActivitySelection={toggleActivitySelection} />}
-          <button className="interactive-wrapper" onMouseEnter={toggleShowTooltip} onMouseLeave={toggleShowTooltip} tabIndex={-1} type="button"><h2>{imageTagForClassification(activity_classification.key)}<span>{name}</span></h2></button>
+          <Tooltip
+            tooltipText={tooltipContent}
+            tooltipTriggerText={<h2>{imageTagForClassification(activity_classification.key)}<span>{name}</span></h2>}
+          />
         </div>
         <div className="buttons-wrapper">
           {nonToggleButtons}
@@ -316,7 +320,7 @@ const ActivityRow = ({ activity, isSelected, toggleActivitySelection, showCheckb
           <ActivityRowTopics
             hasConcept={!!activity_category_name}
             inExpandedView={isExpanded}
-            maxAllowedLength={calculateMaxAllowedLengthForTopicSection({ activity_classification, activity_category_name, readability_grade_level, standard_level_name})}
+            maxAllowedLength={calculateMaxAllowedLengthForTopicSection({ activity_classification, })}
             onTertiaryLine={false}
             topics={topics}
           />
