@@ -10,11 +10,11 @@ interface GradeLevelFiltersProps {
   handleGradeLevelFilterChange: (gradeLevelFilters: number[]) => void,
 }
 
-const GRADE_LEVEL_LABELS = ['2-3', '4-5', '6-7', '8-9', '10-12']
+const GRADE_LEVEL_LABELS = ['4-5', '6-7', '8-9', '10-12']
 const MIN_LEVEL = 0
 const MAX_LEVEL = GRADE_LEVEL_LABELS.length - 1
 
-const tooltipText = "The grade level helps you see which activities are accessible to your students, considering things like readability and the topic of the activity. Students never see the grade level of the activities they’re assigned.<br/><br/>We recommend filtering by the grade you teach, and then using the readability and the concepts filters to narrow down which activities are right for your students. Grades are always presented as a range, rather than as a level, because students in later grades can also benefit from more basic activities.<br/><br/>Click the “?” icon to learn more about how we determine grade ranges and for helpful filtering tips."
+const tooltipText = "The grade level range helps you see which activities are accessible to your students, considering things like readability and the topic of the activity. Students never see the grade level range of the activities they’re assigned.<br/><br/>We recommend filtering by the grade you teach, and then using the readability and the concepts filters to narrow down which activities are right for your students. Grades are always presented as a range, rather than as a level, because students in later grades can also benefit from more basic activities.<br/><br/>Click the “?” icon to learn more about how we determine grade ranges and for helpful filtering tips."
 
 const GradeLevelFilters = ({ gradeLevelFilters, handleGradeLevelFilterChange, }: GradeLevelFiltersProps) => {
   function onChange(valuesArray: number[]) {
@@ -45,13 +45,13 @@ const GradeLevelFilters = ({ gradeLevelFilters, handleGradeLevelFilterChange, }:
       <div className="hoverbox" />
       <div className="name-and-clear-wrapper">
         <h2>
-          <span>Grade Level</span>
+          <span>Grade Level Range</span>
           <a className="focus-on-light interactive-wrapper" href="https://support.quill.org/en/articles/1804176-ccss-vs-readability-how-does-quill-sort-activities" rel="noopener noreferrer" target="_blank"><img alt={helpIcon.alt} src={helpIcon.src} /></a>
         </h2>
         {clearButton}
       </div>
       <div className="slider-wrapper">
-        <label htmlFor="grade-level-slider">Grade Level: {NumberSuffixBuilder(lowestGrade)} - {NumberSuffixBuilder(highestGrade)}</label>
+        <label htmlFor="grade-level-slider">Grade Level Range: {NumberSuffixBuilder(lowestGrade)} - {NumberSuffixBuilder(highestGrade)}</label>
         <TwoThumbSlider
           handleChange={onChange}
           id="grade-level-slider"
